Make ShibbolethMetadataScope more spec compliant
As outlined in the documentation[1] literal and regexp scopes are
not different types of scopes, but rather both are a scope, where
regexp is an attribute of the scope. This attribute is by default
false. In order to better align the implementation with the spec
the constructor is made public and regexp behaviour is now
implemented as an attribute of the scope, rather than as a
different scope altogether.

[1]: https://wiki.shibboleth.net/confluence/display/SC/ShibMetaExt+V1.0