ShibMdScope pattern delimiters are not part of metadata.
According to https://wiki.shibboleth.net/confluence/display/SC/ShibMetaExt+V1.0,
in case of regexp being true, the value of ShibMd:Scope will be solely a
regexp pattern, not including any delimiters. The code should therefore
wrap this before passing it to RegularExpression.

This is a good thing because passing it as a complete regular expression
also allows one to set all modifiers, which is likely undesirable and
a security risk (e.g. by passing the "e" modifier).