Restore error messages on callable assertions
Commit 21b789b switched the isCallable assertions from a custom
implementation to the one from the beberlei assert library. It did not
take into account the signature change of the method.

The signature was: callable value, property path, error message
The signature changed to: callable value, error message, property path

This commit updates all uses of isCallable to the new
signature. Without this fix, all error messages are 'predicate'
instead of the default error message.